# docs/openapi.yaml
openapi: 3.0.0
info:
  title: PROYECTO BASE BACKEND
  version: '1.0'
  description: Documentación del Proyecto base BACKEND
  contact:
    name: agetic@agetic.gob.bo
servers:
  - url: 'http://olimpiadas-backend.test.gtic.gob.bo/olimpiadas-backend'  
paths:
...
  /api/auth/estudiante:
    post:
      operationId: AuthenticationController_loginEstudiante
      summary: Login para estudiantes
      tags: ['autenticacion']
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                usuario:
                  type: string
                  description: Nro. documento de identidad
                  example: Nzg5NDU0NQ
                rude:
                  type: string
                  description: Registro único de estudiante
                  example: UlVERTAwMDE
              required:
                - usuario
                - rude
      responses:
        "200":
          description: Inicio de sesión correcto
          content:
            application/json:
              schema:
                type: object
                properties:
                  finalizado:
                    type: boolean
                    example: true
                    description: Estado de respuesta
                  mensaje:
                    type: string
                    example: Sessión iniciada correctamente
                    description: Mensaje de respuesta
                  datos:
                    type: object
                    example:
                      access_token: eyJhbGciOiJIUzI1Ni...
                      id: 19eacbcc-3519-45ce
                      rude: RUDE0001
                      persona:
                        nroDocumento: "1234567"
                        nombres: JUAN
                        primerApellido: PEREZ
                        segundoApellido: null
                    description: Datos de la respuesta
...
  /api/examenes/{id}/iniciar:
    patch:
      summary: Api para iniciar un examen habilitado
      tags: ['examen']
      operationId: patch-api-examen-iniciar
      description: Permite iniciar un examen para un estudiante, el examen debe estar habilitado 
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: 'aadacc5e-c9ec-4ff8-9ee2-847e4ed502f8'
          description: identificador del examen de un estudiante
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  idExamen: 
                    type: string
                    example: 'aadacc5e-c9ec-4ff8-9ee2-847e4ed502f8'
                  fechaHoraInicio:
                    type: integer
                    example: 1619462591762
                  duracionMinutos:
                    type: integer
                    example: 60
                  tipoPrueba:
                    type: string
                    example: 'ONLINE'
                  estado:
                    type: string
                    example: 'EN_PROCESO'
                  idEstudiante:
                    type: string
                    example: '954c7fdb-f54d-47ef-97d0-10464b244652'
                  area:
                    type: string
                    example: 'Fisica'
                  etapa:
                    type: string
                    example: 'Dos'
                  gradoEscolar:
                    type: string
                    example: 'Segundo de secundaria'
                  preguntas:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: 'd5de12df-3cc3-5a58-a742-be24030482d8'
                        tipoPregunta:
                          type: string 
                          example: 'SELECCION'
                        textoPregunta:
                          type: string 
                          example: 'En un supermercado hay 6 hileras de 6 carros cada una ¿Cuántos carros hay en total?'
                        imagenPregunta:
                          type: string 
                          example: null
                        opciones:
                          type: object 
                          example: { a: 50, b: 100, c: 'Ninguno de los anteriores' }
                        respuestas:
                          type: array 
                          example: null
  /api/examenes/{id}:
    get:
      summary: Api para recuperar un examen en proceso
      tags: ['examen']
      operationId: patch-api-examen-recuperar
      description: Permite recuperar el cuerpo de un examen de un estudiante, este examen debe estar en curso
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: 'aadacc5e-c9ec-4ff8-9ee2-847e4ed502f8'
          description: identificador del examen de un estudiante
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  idExamen: 
                    type: string
                    example: 'aadacc5e-c9ec-4ff8-9ee2-847e4ed502f8'
                  fechaHoraInicio:
                    type: integer
                    example: 1619462591762
                  duracionMinutos:
                    type: integer
                    example: 60
                  tipoPrueba:
                    type: string
                    example: 'ONLINE'
                  estado:
                    type: string
                    example: 'EN_PROCESO'
                  idEstudiante:
                    type: string
                    example: '954c7fdb-f54d-47ef-97d0-10464b244652'
                  area:
                    type: string
                    example: 'Fisica'
                  etapa:
                    type: string
                    example: 'Dos'
                  gradoEscolar:
                    type: string
                    example: 'Segundo de secundaria'
                  preguntas:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: 'd5de12df-3cc3-5a58-a742-be24030482d8'
                        tipoPregunta:
                          type: string 
                          example: 'SELECCION'
                        textoPregunta:
                          type: string 
                          example: 'En un supermercado hay 6 hileras de 6 carros cada una ¿Cuántos carros hay en total?'
                        imagenPregunta:
                          type: string 
                          example: null
                        opciones:
                          type: object 
                          example: { a: 50, b: 100, c: 'Ninguno de los anteriores' }
                        respuestas:
                          type: array 
                          example: null
  /api/examenes/{id}/finalizar:
    patch:
      summary: Api para finalizar un examen en proceso
      tags: ['examen']
      operationId: patch-api-examen-finalizar
      description: Permite finalizar el examen en curso de un estudiante
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: 'aadacc5e-c9ec-4ff8-9ee2-847e4ed502f8'
          description: identificador del examen de un estudiante
      responses:
        '200':
          description: OK
  "/api/examenes/detalle/{id}":
    patch:
      operationId: ExamenController_guardarExamenDetallePorId
      summary: Guardar respuestas de una pregunta
      tags: ['examen']
      parameters:
        - name: id
          required: true
          in: path
          description: Identificador de estudiante-examen-detalle
          example: 9c382541-0120-4045-a98d-63cb6831fb47
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                respuestas:
                  description: Respuestas de la pregunta
                  example:
                    - {}
                  type: array
                  items:
                    type: string
              required:
                - respuestas
      responses:
        "201":
          description: Respuestas guardadas correctamente
          content:
            application/json:
              schema:
                type: object
  "/api/estudiantes/{id}/examenes":
    get:
      operationId: EstudianteController_obtenerEstudianteExamenes
      summary: Obtener examenes disponibles del estudiante
      tags: ["estudiantes"]
      parameters:
        - name: id
          required: true
          in: path
          description: Identificador de estudiante
          example: 19eacbcc-3519-45ce-9de7-1ac2eee0aab7
          schema:
            type: string
      responses:
        "200":
          description: Datos obtenidos correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  finalizado:
                    type: boolean
                  mensaje:
                    type: string
                    example: Datos obtenidos correctamente
                  datos:
                    type: object
                    example:
                      - id: 9fbba8ae-d856-42bd-a7b5-fd91f715db3b
                        estado: ACTIVO
                        calendarioFechaHoraInicio: 1618401600000
                        calendarioFechaHoraFin: 1618405200000
                        areaNombre: Matemáticas
                        etapaNombre: Uno
        "401":
          description: No autorizado
        "403":
          description: Prohibido
...